backend/requirements-dev.txt
# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
hypothesis==6.92.0
moto==4.2.0

# Code quality
black==23.12.0
mypy==1.7.1
flake8==6.1.0
isort==5.13.0

# Type stubs
boto3-stubs[essential]==1.34.0
types-requests==2.31.0
